[in_app_purchase] Update to the latest pkg:json_serializable (#4434)
- Remove unneeded class now that enum helpers can be generated without one
- Code reorganization
- Remove superflous JsonKey values
