--- packages/in_app_purchase/in_app_purchase/build.yaml
@@ -1,7 +1,0 @@
-targets:
-  $default:
-    builders:
-      json_serializable:
-        options:
-          any_map: true
-          create_to_json: true
