+++ packages/in_app_purchase/in_app_purchase_android/pubspec.yaml
@@ -2,10 +2,10 @@ name: in_app_purchase_android
 description: An implementation for the Android platform of the Flutter `in_app_purchase` plugin. This uses the Android BillingClient APIs.
 repository: https://github.com/flutter/plugins/tree/master/packages/in_app_purchase/in_app_purchase_android
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+in_app_purchase%22
-version: 0.1.5+1
+version: 0.1.6
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
   flutter: ">=2.0.0"
 
 flutter:
@@ -21,12 +21,12 @@ dependencies:
   flutter:
     sdk: flutter
   in_app_purchase_platform_interface: ^1.1.0
-  json_annotation: ^4.0.1
+  json_annotation: ^4.3.0
   meta: ^1.3.0
 
 dev_dependencies:
   build_runner: ^2.0.0
   flutter_test:
     sdk: flutter
-  json_serializable: ^5.0.2
+  json_serializable: ^6.0.0
   test: ^1.16.0
