+++ packages/in_app_purchase/in_app_purchase_ios/build.yaml
@@ -1,7 +1,8 @@
+# See https://pub.dev/packages/build_config
 targets:
   $default:
     builders:
       json_serializable:
         options:
           any_map: true
-          create_to_json: true
+          create_to_json: false
