+++ packages/in_app_purchase/in_app_purchase_ios/lib/src/store_kit_wrappers/enum_converters.dart
@@ -23,7 +23,7 @@ class SKTransactionStatusConverter
     if (json == null) {
       return SKPaymentTransactionStateWrapper.unspecified;
     }
-    return _$enumDecode<SKPaymentTransactionStateWrapper, dynamic>(
+    return $enumDecode<SKPaymentTransactionStateWrapper, dynamic>(
         _$SKPaymentTransactionStateWrapperEnumMap
             .cast<SKPaymentTransactionStateWrapper, dynamic>(),
         json);
@@ -64,7 +64,7 @@ class SKSubscriptionPeriodUnitConverter
     if (json == null) {
       return SKSubscriptionPeriodUnit.day;
     }
-    return _$enumDecode<SKSubscriptionPeriodUnit, dynamic>(
+    return $enumDecode<SKSubscriptionPeriodUnit, dynamic>(
         _$SKSubscriptionPeriodUnitEnumMap
             .cast<SKSubscriptionPeriodUnit, dynamic>(),
         json);
@@ -89,7 +89,7 @@ class SKProductDiscountPaymentModeConverter
     if (json == null) {
       return SKProductDiscountPaymentMode.payAsYouGo;
     }
-    return _$enumDecode<SKProductDiscountPaymentMode, dynamic>(
+    return $enumDecode<SKProductDiscountPaymentMode, dynamic>(
         _$SKProductDiscountPaymentModeEnumMap
             .cast<SKProductDiscountPaymentMode, dynamic>(),
         json);
