+++ packages/in_app_purchase/in_app_purchase_ios/pubspec.yaml
@@ -2,7 +2,7 @@ name: in_app_purchase_ios
 description: An implementation for the iOS platform of the Flutter `in_app_purchase` plugin. This uses the iOS StoreKit Framework.
 repository: https://github.com/flutter/plugins/tree/master/packages/in_app_purchase/in_app_purchase_ios
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+in_app_purchase%22
-version: 0.1.3+5
+version: 0.1.4
 
 environment:
   sdk: ">=2.14.0 <3.0.0"
@@ -20,12 +20,12 @@ dependencies:
   flutter:
     sdk: flutter
   in_app_purchase_platform_interface: ^1.1.0
-  json_annotation: ^4.0.1
+  json_annotation: ^4.3.0
   meta: ^1.3.0
 
 dev_dependencies:
   build_runner: ^2.0.0
   flutter_test:
     sdk: flutter
-  json_serializable: ^5.0.2
+  json_serializable: ^6.0.0
   test: ^1.16.0
